+++ src/main/java/com/onlyoffice/manager/document/DefaultDocumentManager.java
@@ -32,6 +32,7 @@ import lombok.Setter;
 
 import java.io.IOException;
 import java.io.InputStream;
+import java.nio.charset.Charset;
 import java.text.MessageFormat;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -342,6 +343,70 @@ public abstract class DefaultDocumentManager implements DocumentManager {
                 .getMaxFileSize();
 
         return size > 0 ? size : DEFAULT_MAX_FILE_SIZE;
+    }
+
+    @Override
+    public boolean isForm(final InputStream inputStream) {
+        try {
+            // CHECKSTYLE:OFF
+            byte[] bytes = inputStream.readNBytes(300);
+            // CHECKSTYLE:ON
+            String pBuffer = new String(bytes, Charset.forName("Windows-1252"));
+
+            int indexFirst = pBuffer.indexOf("%\315\312\322\251\015");
+            if (indexFirst == -1) {
+                return false;
+            }
+
+            // CHECKSTYLE:OFF
+            String pFirst = pBuffer.substring(indexFirst + 6);
+            // CHECKSTYLE:ON
+            if (!pFirst.startsWith("1 0 obj\012<<\012")) {
+                return false;
+            }
+
+            // CHECKSTYLE:OFF
+            pFirst = pFirst.substring(11);
+            // CHECKSTYLE:ON
+
+            int indexStream = pFirst.indexOf("stream\015\012");
+            int indexMeta = pFirst.indexOf("ONLYOFFICEFORM");
+
+            if (indexStream == -1 || indexMeta == -1 || indexStream < indexMeta) {
+                return false;
+            }
+
+            String pMeta = pFirst.substring(indexMeta);
+            // CHECKSTYLE:OFF
+            pMeta = pMeta.substring("ONLYOFFICEFORM".length() + 3);
+            // CHECKSTYLE:ON
+
+            int indexMetaLast = pMeta.indexOf(" ");
+
+            if (indexMetaLast == -1) {
+                return false;
+            }
+
+            pMeta = pMeta.substring(indexMetaLast + 1);
+
+            indexMetaLast = pMeta.indexOf(" ");
+
+            if (indexMetaLast == -1) {
+                return false;
+            }
+
+            return true;
+        } catch (IOException e) {
+            return false;
+        } finally {
+            if (inputStream != null) {
+                try {
+                    inputStream.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
     }
 
     protected static void init() {